perf: optimize order lookup query (4.1s -> 133ms)

+++ ordering/src/main/java/com/careup/ordering/domain/order/dto/response/AllBranchesSalesResponseDto.java
@@ -16,7 +16,7 @@ public class AllBranchesSalesResponseDto {
     private String periodType;
     private Long totalSales; // 전체 기간 총 매출
     private Long totalOrders; // 전체 기간 총 주문
-    private Integer totalBranchCount; // 총 지점 수
+    private Long totalBranchCount; // 총 지점 수
     private List<AllBranchesSalesDto> salesData; // 기간별 매출 데이터
 
     // 전일 대비 증가율 (%)
